check credentials for every database request

+++ src/main/java/ru/students/lab/controllers/CollectionController.java
@@ -1,9 +1,12 @@
 package ru.students.lab.controllers;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import ru.students.lab.database.CollectionModel;
 import ru.students.lab.database.Credentials;
 import ru.students.lab.database.SQLQuery;
 import ru.students.lab.database.UserModel;
+import ru.students.lab.exceptions.AuthorizationException;
 import ru.students.lab.models.Dragon;
 
 import java.security.NoSuchAlgorithmException;
@@ -13,6 +16,7 @@ import java.util.concurrent.locks.ReentrantLock;
 
 public class CollectionController {
 
+    protected static final Logger LOG = LogManager.getLogger(CollectionController.class);
     private final CollectionModel collectionModel;
     private final UserModel userModel;
     private final ReentrantLock mainLock;
@@ -38,7 +42,7 @@ public class CollectionController {
             else
                 return "User/Password given not found or incorrect";
         } catch (SQLException | NoSuchAlgorithmException ex) {
-            ex.printStackTrace();
+            LOG.error("logging in", ex);
             return ex.getMessage();
         }
     }
@@ -51,50 +55,74 @@ public class CollectionController {
             else
                 return credentials;
         } catch (Throwable ex) {
+            LOG.error("registering user", ex);
             return ex.getMessage();
         }
     }
 
     public String addDragon(int key, Dragon dragon, Credentials credentials) {
         try {
+            if (assertUserNotExist(credentials))
+                throw new AuthorizationException();
+
             return collectionModel.insert(key, dragon, credentials);
         } catch (Throwable ex) {
+            LOG.error("inserting dragon in db", ex);
             return ex.getMessage();
         }
     }
 
     public String updateDragon(int id, Dragon dragon, Credentials credentials) {
         try {
+            if (assertUserNotExist(credentials))
+                throw new AuthorizationException();
+
             return collectionModel.update(id, dragon, credentials);
         } catch (Throwable ex) {
-            ex.printStackTrace();
+            LOG.error("updating dragon in db", ex);
             return ex.getMessage();
         }
     }
 
     public String deleteAllDragons(Credentials credentials) {
         try {
+            if (assertUserNotExist(credentials))
+                throw new AuthorizationException();
+
             return collectionModel.deleteAll(credentials);
         } catch (Throwable ex) {
-            ex.printStackTrace();
+            LOG.error("deleting all dragons in db", ex);
             return ex.getMessage();
         }
     }
 
     public String deleteDragon(int key, Credentials credentials) {
         try {
+            if (assertUserNotExist(credentials))
+                throw new AuthorizationException();
+
             return collectionModel.delete(key, credentials);
         } catch (Throwable ex) {
-            ex.printStackTrace();
+            LOG.error("deleting dragon in db", ex);
             return ex.getMessage();
         }
     }
 
-    public int[] deleteDragonsGreaterThanKey(int key, Credentials credentials) throws SQLException {
+    public int[] deleteDragonsGreaterThanKey(int key, Credentials credentials) throws SQLException, NoSuchAlgorithmException {
+        if (assertUserNotExist(credentials))
+            throw new AuthorizationException();
+
         return collectionModel.deleteOnKey(key, credentials, SQLQuery.Delete.DRAGONS_WITH_GREATER_KEY);
     }
 
-    public int[] deleteDragonsLowerThanKey(int key, Credentials credentials) throws SQLException {
+    public int[] deleteDragonsLowerThanKey(int key, Credentials credentials) throws SQLException, NoSuchAlgorithmException {
+        if (assertUserNotExist(credentials))
+            throw new AuthorizationException();
+
         return collectionModel.deleteOnKey(key, credentials, SQLQuery.Delete.DRAGONS_WITH_LOWER_KEY);
+    }
+
+    public boolean assertUserNotExist(Credentials credentials) throws SQLException, NoSuchAlgorithmException {
+        return userModel.checkUserAndGetID(credentials) == -1;
     }
 }
